% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPrep.R
\docType{package}
\name{DEPrep}
\alias{DEPrep}
\alias{DEPrep-package}
\title{DEPrep: A package for preparing gene count data for differential
expression analysis}
\description{
This package provides means to manipulate, filter and convert gene count data
in preperation for differential expression
}
\author{
Louise Whiteside \email{lewhiteside@gmail.com}
}
