% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffprep.R
\name{diffprep}
\alias{diffprep}
\title{DEPrep}
\usage{
diffprep(data, gene_col_name, filter_number)
}
\arguments{
\item{data}{A \code{data.frame} containing gene counts}

\item{gene_col_name}{the column name of genes in original dataframe}

\item{filter_number}{the number threshold of gene counts to which the data are filtered}
}
\value{
A filtered \code{matrix} with genes as the row names
}
\description{
Prepares dataframe of gene counts for differential expression analysis.
}
\examples{
\dontrun{
diffprep(genecounts, "gene_names", 5
}

}
\author{
Louise Whiteside
}
