% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genecounts}
\alias{genecounts}
\title{DiffPrepare example data set}
\format{
A data frame of integers with 10 rows and 3 variables
\describe{
\item{gene_names}{gene IDs}
\item{treatment1}{counts within treatment 1}
\item{treatment2}{counts within treatment 2}
\item{treatment3}{counts within treatment 3}
}
}
\source{
Randomly generated data
}
\usage{
genecounts
}
\description{
I have made some example data for the package
}
\keyword{datasets}
